feat: create a library to manage subscriptions
This library allows managing the customer
subscriptions that are stored in the customerdataapi
microservice.

Provides a class called UpgradeSubscription
for upgrading the subscription level in the
configuration data of a specific customer.

It also provides a class called DowngradeSubscription
for downgrading the subscription level in the
configuration data of a specific customer.