ci: add ci files for subscription-manager library
Creates 2 CI files, one for testing and the other for QA.

These files are separated because handling a job per file
keeps files smaller and more manageable.

- test_subscription_manager_lib.yml: Contains the
'test' job that will run all the unit tests of the subscription
manager library and display the test coverage, which should
be at least of 90%.

- qa_subscription_manager_lib.yml: Contains the
'code-quality-check' job that has a step for linting
and a step to check the format of the code.