feat: add exit codes to the library
Adds exit codes to all the errors in the library that are
generated with the logging module.

This feature was added as part of the challenge of including
the exit code in the error messages by using bash.

This is an example of joining the exit code with the error message:
'Error code 5: Attempted to downgrade from basic to premium.'

Other changes:
- Adds an 'exit_code' attribute to the SubscriptionManager
class to store the exit code on error.
- Includes all the necessary unit tests.
- Modifies some tests to fit this new feature.
- Modifies the logging module configuration to send the
errors to a file.
- Creates the file error.log that the logging module will
use to write the errors.
- Modifies the cli file to join the exit code with the
error message.

Additional notes:
- The meaning of each exit code will be included later
in the README.md file of the library.